remove deepspeed dependency
